monitor-service/requirements.txt
tensorflow==2.18.0
numpy==1.26.4
Pillow==9.0.1
firebase-admin==5.0.0
websockets==10.3
torch==2.5.1
torchvision==0.20.1


# tensorflow==2.17.1
# numpy==1.24.3
# Pillow==9.2.0
# firebase-admin==5.0.0
# websockets==10.3
# torch==2.0.1
# torchvision==0.15.2